Use Android DownloadManager instead of sending view intent to browser.
DownloadManager is built in to Android, and so will always be available,
even when no web browser is installed on the system yet.

Because I only know English, I have done the best I could with the other
translations. The "Download started" text was from the Microsoft Language
Portal (specifically from Skype), and the infobox text was edited with the
help of Google Translate and some dictionaries.